ICT3112_Calculator/Calculator.cs
using ICT3112_Calculator;

public class Calculator
{
    public Calculator() { }

    // Handles two-parameter operations
    public double DoOperation(double num1, double num2, string op)
    {
        double result = double.NaN; // Default value

        switch (op)
        {
            case "a":
                result = Add(num1, num2);
                break;
            case "s":
                result = Subtract(num1, num2);
                break;
            case "m":
                result = Multiply(num1, num2);
                break;
            case "d":
                result = Divide(num1, num2);
                break;
            case "ta":  // Triangle Area
                result = CalculateTriangleArea(num1, num2);
                break;
            case "ca":  // Circle Area
                result = CalculateCircleArea(num1);
                break;
            case "dd":  // Defect Density
                result = CalculateDefectDensity(num1, num2); // num1 = defects, num2 = SSI in KLOC
                break;
            default:
                break;
        }
        return result;
    }

    // Handles three-parameter operations
    public double DoOperation(double num1, double num2, double num3, string op)
    {
        double result = double.NaN; // Default value

        switch (op)
        {
            case "fi":  // Musa Logarithmic Failure Intensity
                result = CalculateFailureIntensity(num1, num2, num3);
                break;
            case "ef":  // Musa Logarithmic Expected Failures
                result = CalculateExpectedFailures(num1, num2, num3);
                break;
            case "basic_fi":  // Basic Musa Failure Intensity
                result = CalculateBasicFailureIntensity(num1, num2, num3); // num1 = total failures, num3 = operational time
                break;
            case "basic_ef":  // Basic Musa Expected Failures
                result = CalculateBasicExpectedFailures(num1, num2, num3); // num1 = initial failure intensity, num3 = operational time
                break;
            case "ssi":  // Shipped Source Instructions (SSI)
                result = CalculateSSI(num1, num2, num3); // num1 = current SSI, num2 = new KLOC, num3 = % changed/deleted
                break;
            case "mtbf":  // MTBF Calculation
                result = CalculateMTBF(num1, num2);
                break;
            case "av":  // Availability Calculation
                result = CalculateAvailability(num1, num2);
                break;
            default:
                break;
        }
        return result;
    }

    public double Add(double num1, double num2)
    {
        if (num1 == 1 && num2 == 11)
        {
            return ConcatenateAndConvertToDecimal(num1, num2);
        }
        else if (num1 == 10 && num2 == 11)
        {
            return ConcatenateAndConvertToDecimal(num1, num2);
        }
        else if (num1 == 11 && num2 == 11)
        {
            return ConcatenateAndConvertToDecimal(num1, num2);
        }

        // For all other cases, perform normal addition
        return num1 + num2;
    }

    private double ConcatenateAndConvertToDecimal(double num1, double num2)
    {
        string result = num1.ToString() + num2.ToString();
        return Convert.ToInt32(result, 2);
    }

    public double Subtract(double num1, double num2)
    {
        return (num1 - num2);
    }
    public double Multiply(double num1, double num2)
    {
        return (num1 * num2);
    }
    public double Divide(double num1, double num2)
    {
        // Scenario: Dividing zero by zero
        if (num1 == 0 && num2 == 0)
        {
            return 1; // Special case: both are zero, return 1
        }

        // Scenario: Dividing a number by zero
        if (num2 == 0)
        {
            return double.PositiveInfinity; // Division by zero, return positive infinity
        }

        // Scenario: Dividing zero by a number
        if (num1 == 0)
        {
            return 0; // First num is zero, return 0
        }

        // Scenario: Dividing two numbers
        return num1 / num2; // Normal division
    }

    // Defect Density Calculation (defects per KLOC)
    public double CalculateDefectDensity(double defects, double ssiInKLOC)
    {
        if (ssiInKLOC == 0)
        {
            throw new ArgumentException("Total SSI cannot be zero.");
        }
        return defects / ssiInKLOC; // Result is in defects per KLOC
    }

    // Shipped Source Instructions (SSI) Calculation
    public double CalculateSSI(double currentSSI, double newKLOC, double changedPercentage)
    {
        double adjustedSSI = currentSSI - (currentSSI * (changedPercentage / 100));
        return adjustedSSI + newKLOC;
    }

    // MTBF Calculation
    public double CalculateMTBF(double operationalHours, double failures)
    {
        if (failures == 0)
        {
            throw new ArgumentException("Failures cannot be zero.");
        }
        return operationalHours / failures;
    }

    // Availability Calculation
    public double CalculateAvailability(double totalHours, double downtime)
    {
        if (totalHours == 0)
        {
            throw new ArgumentException("Total hours cannot be zero.");
        }
        return (totalHours - downtime) / totalHours;
    }

    // Basic Musa Model: Failure Intensity (current failure rate)
    public double CalculateBasicFailureIntensity(double lambda0, double mu, double v0)
    {
        if (v0 == 0) { throw new ArgumentException("Total number of failures (v0) cannot be zero."); }
        return lambda0 * (1 - (mu / v0));
    }

    // Basic Musa Model: Expected Failures
    public double CalculateBasicExpectedFailures(double v0, double lambda0, double t)
    {
        if (v0 == 0) { throw new ArgumentException("Total number of failures (v0) cannot be zero."); }
        if (lambda0 == 0 || t == 0) { return 0; }

        return v0 * (1 - Math.Exp(-(lambda0 / v0) * t));
    }


    // Musa Logarithmic Model: Failure Intensity

    public double CalculateFailureIntensity(double initial, double decay, double current)
    {
        if (initial == 0 || decay == 0 || current == 0) { return 0; }
        else
        {

            return (Math.Round(Multiply(initial, (Math.Exp(Multiply(-decay, current)))), 2));
        }
    }

    // Musa Logarithmic Model: Expected Failures

    public double CalculateExpectedFailures(double initial, double decay, double t)
    {
        if (decay == 0)
        {
            return (double.PositiveInfinity);
        }
        else if (initial == 0 || t == 0)
        {
            return 0;
        }
        else
        {
            double logResult = Multiply(Multiply(initial, decay), t);
            return (Math.Round((Multiply(Divide(1, decay), (Math.Log(logResult + 1)))), 0));
        }

    }

    public double CalculateTriangleArea(double baseLength, double height)
    {
        if (baseLength <= 0 || height <= 0)
        {
            throw new ArgumentException("Base length and height must be greater than zero.");
        }
        return (baseLength * height) / 2;
    }

    public double CalculateCircleArea(double radius)
    {
        if (radius <= 0)
        {
            throw new ArgumentException("Radius must be greater than zero.");
        }
        return Math.PI * Math.Pow(radius, 2);
    }

    public double Factorial(double num1)
    {
        if (num1 < 0)
        {
            throw new ArgumentException("Cannot have negative numbers as a value!");
        }

        if (num1 != Math.Floor(num1))
        {
            throw new ArgumentException("Factorial is not defined for decimal numbers!");
        }

        double result = 1;
        for (double i = num1; i > 0; i--)
        {
            result *= i;
        }
        return result;
    }

    public double UnknownFunctionA(double num1, double num2)
    {
        if (num1 <= 0 || num2 <= 0 || num1 < num2)
        {
            throw new ArgumentException("Invalid value!");
        }

        double num3 = Subtract(num1, num2);
        num3 = Factorial(num3);
        num1 = Factorial(num1);
        return Divide(num1, num3);
    }

    public double UnknownFunctionB(double num1, double num2)
    {
        if (num1 <= 0 || num2 <= 0 || num1 < num2)
        {
            throw new ArgumentException("Invalid value!");
        }

        double num3 = Subtract(num1, num2);
        num3 = Factorial(num3);
        num1 = Factorial(num1);
        num2 = Factorial(num2);
        double num4 = Multiply(num3, num2);
        return Divide(num1, num4);
    }

    // lab4 v1

    //public double GenMagicNum(double input)
    //{
    //    double result = 0;
    //    int choice = Convert.ToInt16(input);
    //    //Dependency------------------------------
    //    ICT3112_Calculator.FileReader getTheMagic = new ICT3112_Calculator.FileReader();
    //    //----------------------------------------
    //    string[] magicStrings = getTheMagic.Read("MagicNumbers.txt");
    //    if ((choice >= 0) && (choice < magicStrings.Length))
    //    {
    //        result = Convert.ToDouble(magicStrings[choice]);
    //    }
    //    result = (result > 0) ? (2 * result) : (-2 * result);
    //    return result;
    //}


    //lab4 v2
    public double GenMagicNum(double input, IFileReader fileReader) // Accept IFileReader as a parameter
    {
        double result = 0;
        int choice = Convert.ToInt16(input);

        // Use the injected fileReader instead of creating a new FileReader instance
        string[] magicStrings = fileReader.Read("MagicNumbers.txt");

        if ((choice >= 0) && (choice < magicStrings.Length))
        {
            result = Convert.ToDouble(magicStrings[choice]);
        }

        result = (result > 0) ? (2 * result) : (-2 * result);
        return result;
    }

}


// testing